requirements.txt
fastapi
uvicorn[standard]
mistralai
numpy
streamlit
pydantic
python-dotenv
httpx
pytest
